gha-cost-predictor/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
pyyaml==6.0.2
python-dotenv==1.0.1
python-multipart==0.0.12
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
apscheduler==3.10.4
